device-type: Do not disable soft keyboard
EventHub detects the built in PS/2 controller as a physical keyboard
and disables the soft keyboard. The input service has been patched
to ignore the specific device, but the "feature" needs to be enabled
by setting ro.ignore_atkbd to not zero.

Enabling the feature for tablet and phone devices.

As we do not want to enable this for laptop style designs (hsb) a new
device type, clamshell is introduced and a corresponding patch to update
the hsb confifuration has been created.

Issue: APDEV-1361

Depends-Change-Id: Ie8e22ef0d3053492faad10040399d6bc1ec9c62b

Change-Id: Ic6e11bf6df735f4403c3e5f6626d3e738926c721
Signed-off-by: Axel Fagerstedt <axel.fagerstedt@intel.com>
Reviewed-on: https://android.intel.com/162191
Reviewed-by: Lilja, Ola <ola.lilja@intel.com>