External SDcard not used as external storage anymore
The goal of an EXTERNAL_STORAGE (both primary or secondary)
is to extend the device internal storage space available for
the applications. According to Android External Storage
Technical Information external storage devices must be a
semi-permanent part of the device (such as an SD card slot
in a battery compartment). From that perspective an SD card
inserted in a laptop SD card reader is never a
semi-permanent part of the device.

Issue: APDEV-934
Issue: APDEV-669
Issue: AXIA-4965
Change-Id: Idde5038761a19f2b82716dd47375e92bddb7f515
Category: Device Enablement
Domain: Storage
Origin: internal
Upstream-Candidate: No
Signed-off-by: Viorel Suman <viorel.suman@intel.com>
Reviewed-on: https://android.intel.com/165350
Reviewed-by: Tomaszewski, Piotr <piotr.tomaszewski@intel.com>
Reviewed-by: Chiras, Robert <robert.chiras@intel.com>
Reviewed-by: Zhang, Jianxun <jianxun.zhang@intel.com>
Reviewed-by: Boie, Andrew P <andrew.p.boie@intel.com>
Tested-by: Boie, Andrew P <andrew.p.boie@intel.com>